+++ Book.js
@@ -1,0 +1,12 @@
+class Book {
+  constructor(id, title, name) {
+    this.id = id;
+    this.title = title;
+    this.name = name;
+    this.node = null;
+  }
+
+    pageimReading = 0
+}
+
+export default Book;
