+++ index.html
@@ -29,6 +29,6 @@
         </form>
     </section>
 
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
